Split app metadata from model metadata
